Build changes in vrouter to support dkms.
With this change, the Makefile in the dkms package will support make
without specifying any arguments (as any stand-alone module makefile
should be) in the system where the package is installed. While building
using 'SCons' (as part of standard build), we will provide arguments
to make so that it understands where to pick some files from, as they
exist in the source tree.