xen: Fix for GRO path
Existing vrouter code implements GRO with the option of using hard header length of the
device for GRO flow merge. This option does not exist in older kernel, where flow comparison
is based on mac header. Fixed it by adjusting the the mac header of the skb for xen kernel.

Signed-off-by: Anirban Chakraborty <abchak@juniper.net>