If the mac address of the packet does not match with the mac address of the physical interface in which the packet came, just drop the packet.
Other than broadcast and multicast macs, the mac address of the packet
will not match with that of the interface typically when the interface goes
into promiscuous mode. When in promiscuous mode, vrouter will forward all
the packets that it receives from the physical interface. This situation
takes an ugly turn when the switch broadcasts unknown unicast packets and
all the computes receives those packets and starts forwarding, resulting
in a storm.

Change-Id: Ib76963e940996e1e4239988f48f57cc69170ea54
Closes-Bug: 1350774