In the pull inner header functions, get the outer ip header pointer with the help of pkt_network_header rather than as an argument from parent functions
When we receive an MPLSoUDP encapsulated packet, we pull till the end of
inner transport header to a contiguous buffer. As part of that we also do
checksum verification, if hardware has not done verification. To verify
checksum, ip header is needed and the header pointer is passed to checksum
verification functions. However, in case the headers are in different buffers
and the pull actually pulls data from different buffers, the ip pointer will
not be valid, since pulls can realloc buffers and hence can cause issues.

Closes Bug: 1352294

Change-Id: I918cd6c3046078abf07981e2438f2d11e914ac8a