Fixes PR 1382402, ICMPv6 Packet Too Big not sent
Set the ipv6 next-header value correctly when responding with
Packet Too Big message. In the absence of this, the response
goes with whatever is the next header value in the received
packet and fails if it is not icmpv6 (TCP, Frag-header etc).
In the ping -s 1500 case mentioned in the PR, the next-header
was frag-header.

Change-Id: If921d4609fe54c576bfe314eebbd08e889530b18
Closes-bug: 1382402