Support for VRF translation nexthops (also known as table nexthops)
A vrf translation nexthop changes the vrf of the packet and forwards
the packet in the new vrf. In the VCPE use case, packets that arrive
from the uplink and those that need to be forwarded to one of the other
physical ports, should be subjected to VRF translation before
forwarding. In this case, a label will point to the translation nexthop.

Also,

Fix compilation warnings for utils/ by making a declarion in vr_interface.h
inside __KERNEL__

Make Scons compile the kernel module always, thus making 'make' do the
decision to build the module or not

Closes-Bug:#1386573
Change-Id: Ibba85eec0edddf34c041c4237cb2b5f93f3d4e67