linux tunnel interface support in vrouter
For tunnel interface support, packets that ingresses and egresses the
tunnel interface will have network header rather than ethernet header
as vrouter today expects. vrouter will look the encap of the device
that is added to vrouter to see whether the interface supports an
ethernet encap or not. For now, if the interface does not support an
ethernet encap, vrouter will assume that the interface is a tunnel
interface.

TODO: segmentation and addition of ethernet header for packets that
go to vhost interface during vrouter reset

Bug:#1391051
Change-Id: I47717f4a9e332b8fcfbb2878c4659a5eead54ffb