Carry 'tome' flag across flow queue enqueue and flush
If a packet is routed and the routing decision points to a nexthop
with policy, the packets that are enqueued in the flow entry has
to be routed (and not bridged) when they are flushed from the entry.
To do that, fmd_to_me flag has to be carried across enqueue and
flush.

Closes BUG: #1406873

Change-Id: Ie6b6d545abe814bfab7e6df35e53b1229984a92c