Do not count packets in both sub-interface and in the parent interface that hosts the sub-interface
There is a requirement that stats be not counted in both the underlying
interface and in the sub-interface (double count). This behavior is a
specific requirement for VM interfaces since the VM interface and the
sub-interface on top are considered two different VMIs and they can also
be in two different VRFs. Counting stats in both interfaces will wrongly
put statistics in a VRF where the statistics should not have been counted.
Also, even if the interfaces are in the same VRF, we can't count them
in the sub-interface and the hosting interface, since that will result in
double counting

Closes BUG: #1415427

Change-Id: Icb6485365391ea663ba89b8f1225f093993ee86f