Do not apply rewrite information, if the packet is from vhost and destined to fabric interface. Instead just retain the original layer 2 header (if one is present).
In almost all cases, traffic from vhost will have the correct mac, since arp
resolutions of the host will be correct. Exceptions to this behavior happens
when vrouter is expected to do real routing on the traffic (as opposed to
superficial routing that vrouter does for traffic from vhost and going out on
fabric). Examples of such exceptions are the traffic to local vms over link local
IPs. The superficial routing causes issues when the routing table entries in the
host and vrouter differs. For e.g.: host can add a different gateway for a specific
network, while vrouter has only one default gateway. In this case, though the host
sends traffic to the correct mac (because of proper ARP resolution), vrouter sends
it to default gateway. In such cases, vrouter should not apply rewrite information,
while picking all the other information (such as the egress device).

Closes BUG: #1416619

Change-Id: Ic30b026a852430271202ac03a9ce588d96deb3e4