Move DHCP flood flag to bridge entry
There are cases where a different DHCP server is desired than the default
DHCP server provided by agent. In such cases, a control is needed in vrouter
to not trap the packet to agent. Currently, that control is in interface. The
control can't be in an interface since there are cases where interface will
not be present in vrouter, but still DHCP flood is needed. So, a better place
is to move it to the bridge table entry.

Closes BUG: #1417067

Change-Id: I810874c4d3c137f016930682d76525d656175a61