Allocate hold array when the action is set as 'hold' in the flow table entry
HOLD is an action that is allowed to be set in a flow entry. Agent can set HOLD
as action and expect the first packet to be trapped. To hold packets in an
entry there needs to be hold list (an array, rather). While this hold list is
allocated when a new flow is created by kernel, the hold list is freed once all
the cached packets are flushed. A subsequent 'HOLD' set needs to have the hold
list allocated so that packets can be cached.

Closes BUG: #1425992

Change-Id: Ic32a03f402278a351c72cb6a4f72bafdaad2149c