vrf for packets sent to agent via encap nexthop should come from metadata and not from ingress interface
Till now, packets that were sent to agent via an encap nexthop, had the vrf
in the agent header set to the ingress interface vrf. The vrf value in all
cases have to be the vrf to which the packet has been classified to.

Bug fix for arp behavior: If an arp request ingressed from fabric interface
and if the proxy bit was set in the route, but the stitching information was
not present, vrouter used to proxy with the vhost mac. This behavior is wrong.
vrouter should proxy only if the destination is hosted by it or if the
destination is not hosted and it is a TSN or if the request was meant for a
host in the cp port.

Closes-Bug: #1428135

Change-Id: Ibe543c743a07834a87e1512ac305f1e0f32d5d30