Add filter support for dropstats utility
The dropstats command line utility now provides a means to register
and unregister a filter. A filter is a tuple consisting of source ip,
destination ip, source port, destination port and a vrf id. Each
component of the tuple is optional, and if not provided is treated like
a wild card. The existing functionality of global counters works as
before. To query the filtered stats, a flag needs to be passed to the
dropstats utility.

New sandesh message is  added to register, unregister, and query the
filter.

The vrouter kernel module processes the sandesh message to register,
unregister and query the filter. If a filter is set, for each packet on
the tx and rx side it sets a flag in vr_packet to remember that the packet
matches the filter. Later if the packet is dropped, the filter specific
stats are updated along with the global stats.

Limitations of the current implemetation are 1)only supports ipv4 overlays, with TCP and UDP. 2) works only in linux.

Change-Id: If054186761c78a47484cb5dea222b095dc7aefe8
Implements: blueprint filter-based-dropstats