Validate vif to device and device to vif linkage before using vif
Once rps is done, the handler tries to extract vif from the control
block. From that vif we try to get the dev and the rx_handler_data
(which should be equal to the vif). From whatever information that
was present in the trace, it was confirmed that rx_handler_data was
null. There could be a case where agent restarts and is in the process
of soft resetting the vrouter when this condition can happen. Since
we do not have dump, for now we will check the linkages between vif
and device (<=>) before proceeding with packet processing.

Closes Bug: #1430279

Change-Id: I98066588c3c0d95f1327d505495f2058d9616a4d