Do a route table lookup post vrf translation
In the most common packet path, a route lookup always happens after
flow lookup. But there can be cases when a flow lookup happens post
a route table lookup. In such cases, if a translation happens, then
the destination ip should be looked up in the new route table.
Exception to this case is when packets come from fabric where the
label already points to the destination nexthop and no further lookup
is needed (which is what happens today).

Change-Id: If2191dc73b9764107900119cffa4eabbfc23c899
Closes-Bug: #1435746