Checksum offloading support.
vRouter advertises checksum offloading capabilities to VM. If VM wants to use
the offloading, proper flag in packets is set. vRouter checks for the flag and
either computes checksum in software or offloads to the NIC installed in the
compute node (if NIC supports hardware checksum computing).

The code contains functions ported from DPDK 1.8.0.

Change-Id: I9faa9f06c499911e69ca88a53cc71d082ea53986