Free the defer data in case of errors
To make sure that we flush all the packets that are queued in a flow
entry, we run a defer function. If for any reason this defer was not
scheduled (because the function was called with no hold queue), the
defer data has to be freed.

Closes-BUG: #1436798
(cherry picked from commit 8c30ce9a3e254d45dc4de595cc066f2da21c18d6)

Conflicts:
	dp-core/vr_flow.c
Change-Id: Ideabe252ff6c56c6e01ccbc3a27044bac09933f3