Increase MPLS table size to 128K
Also, avoid overflows and release the reference count of nexthops in
case of a change to an existing label.

Change-Id: I9fb14dbf58ad0c8b0123614c29388eaacf5e069f
Partial-BUG: #1446170

Increase the number of nexthops that a vrouter can host

The existing limit for the number of nexthops in a compute node is
64k. To accommodate a bigger number of nexthops, the nexthop type
has to change to integer. Also, the nexthop table type has to change
from a simple array to big table so that a larger table can be created.

The number of mpls labels and the number of nexthops are made module
load parameters so that the memory demand is made only on nodes where
there is a real requirement.

Change-Id: I42ff27ab3b92470852b851d161700d4e35c5e489
Partial-BUG: #1446170

For non tcp/udp/icmp/sctp protocols, port values used in the
five tuple of the flow key should be 0

Change-Id: Ie165da334f5f6e4126bc375a9d3652d0bda6cff9
Closes-BUG: #1447558

Export vRouter table limits and build information to agent/utilities

All the various vRouter limits such as number of interfaces, nexthops,
mpls labels etc.: are now exported in the vrouter_ops sandesh. A new
command ('vrouter') is introduced.

vrouter --info

will display build information and all the vRouter limits.

Change-Id: Ibd367ef7705987f738bead25c236792165cd3240
Partial-BUG: #1446170

Cherry-pick conflicts
Partial-BUG: #1446170

Change-Id: I08d6740d0690ca885f3a28b7e3e8ddac05474974