Increase the number of nexthops that a vrouter can host
The existing limit for the number of nexthops in a compute node is
64k. To accommodate a bigger number of nexthops, the nexthop type
has to change to integer. Also, the nexthop table type has to change
from a simple array to big table so that a larger table can be created.

The number of mpls labels and the number of nexthops are made module
load parameters so that the memory demand is made only on nodes where
there is a real requirement.

Change-Id: I42ff27ab3b92470852b851d161700d4e35c5e489
Partial-BUG: #1446170