Make interface notifier to act only in default namespace
vRouter listens for netdevice register/unregister events so that it
can cleanup its view of interfaces that it uses. The way vRouter
identifies an interface is through the name of the interface.

When docker is run, docker also can create interfaces with the same
name and interface index, albeit in a different name space. In this
particular case, docker creates a veth pair, assigns name of "eth0"
to one part of the pair (that is used by the container) and at exit
deletes the interface. vRouter wrongly assumes that the "eth0" delete
is really for the "eth0" in the default name space and detaches the
interface from vRouter.

Change-Id: I2ba384ba2b85f19f8e666d858da0ff443dacc7ec
Closes-BUG: #1449919