NULL check for the presence of nexthop table
'exit' of the nexthop module will try to access all nexthops in the
nexthop table and release them. If the nexthop table allocation
fails (specifically when crash kernel is booted) during 'init', an
exit will be called and a crash can happen. As is consistent with all
the other exits, a check is needed to make sure that the exit works
regardless of what had happened in the init.

Change-Id: I0baf03e452fc713fe0dabc2c700e5039c7f96fa3
Closes-BUG: #1446170