DPDK ring/interface counters.
DPDK introduces macros to count sent/received packets and their drops
in rte_port library. Those macros updates proper structures containing
variables storing those values. They could be accessed from a DPDK
application.

This commit adds support of those functionality in vRouter. Moreover,
it introduces similar macros and structures for KNI and VIRTIO interfaces.

Counted packets are displayed in the vif utility in DPDK mode as:
- Packets sent to vif
- Packets enqueued on TX ring
- Packets received on vif
- Packets enqueued on RX ring
and, additionaly, drops for each of this counters.

Furthermore, from now on, vif utility allows to check statistics for separate,
CPU core, as well as summed up for all cores (as it was before). This feature
is accessible in both kernel and DPDK modes. Example:
    $ vif --list --core 3  # all interfaces, CPU core no.  3
    $ vif --get 3 -C 4     # interface no. 3, CPU core no. 4

Change-Id: Idab7b02bd7bd058f8da33f44f6826f8b14d4e4f7
Closes-Bug: #1453777