Pull a transport header only if one is present
When packets arrive from a vm/fabric, we try to pull all data till
the first 8 bytes of a transport header into the first buffer so
that linear access to data is possible (keys to flow is what we
look for in the transport header). We do this operation
without checking whether the packet is a fragment or not and such
an unconditional attempt at pull can result in pull failures for
fragments whose data length is less that 8.

Hence, pull only for packets that have a valid transport header and
that has a trapsort protocol we recognize.

Change-Id: Iaf8ec480bef045c774630a7c0cc9afbc867a6062
Closes-BUG: #1460218