TCP connection state awareness for faster flow aging
If a flow has already seen a tcp session close (either through the FIN
mechanism or through the RST mechanism), that flow can be dismantled
immediately instead of waiting for the flow aging time. vRouter will
now track the connection closure (RST as well as FIN/FIN-ACK/ACK or
FIN-ACK-FIN-ACK) and when it comprehends a session close will send a
trap to agent to indicate that the flow(forward/reverse) can be
dismantled.

Similarly, connections to non-existent systems can result in inactive
flows that could be dismantled after proper backouts. vRouter will track
SYN, SYN-ACK, ACK sequence to mark the flow with flags indicating that
it has seen SYN and a session establishment (if the cycle goes through),
which can be used to dismantle dummy flows.

Change-Id: I343fb6d56ef16a08d1a9dabe1027d3426249a17a
Partial-BUG: #1362701