For UUC fllod, treat L2 bridge entry with Drop NH as L2 miss
For DHCP trapping purpose an L2 bridge entry is added in VRF
with discard nexthop. But when Unknown unicast flood feature
is enabled for a VNI, because of the above route, we will have
bridge hit, and UUCF does not get kicked-off. To fix this,
an L2 bridge entry with drop nexthop is treated as L2 miss.

Change-Id: Ib0b6e02bf10214aa1d9a7b56d63df8bcd4583f24
closes-bug: #1466584