Disabling Source lookup for Vxlan packets
When the Vxlan packet from TOR are reaching a compute node,
before processing them in Vrf translate nexthop, we verify,
in nh_output(), whether flow processing needs to happen or not. IF
policy bit is not enabled in the nexthop, we do inner source IP lookup
to identify if it is ECMP source. If so, we force the flow processing.
Incase of control node peering with two MX, the BMS subnet would be
advertised by both MX leading to ECMP source . This is forcing the vxlan
tagged packets to be trapped to AGent with nh_id as VRF translate
nexthop. This leads to Agent marking the flow as short flow as Agent
already has set this flow with l3_nhid.
To avoid this scenario, the source lookup is done only for non vxlan
packets. This check of verifyin whether this is a vxlan packet or not
enabled under a proc entry vxlan_ecmp_no_rpf and by default this is
turned on, which means we dont do RPF check for VXlan packets. If
someone wants to do RPF check even for Vxlan tagged packets, they can
set this to 0.

Change-Id: I4446a0b8171951f206cc13e12855be506b73e53d
closes-bug: #1467273