Disabling Source lookup for Vxlan packets
When the Vxlan packet from TOR are reaching a compute node,
before processing them in Vrf translate nexthop, we verify,
in nh_output(), whether flow processing needs to happen or not. IF
policy bit is not enabled in the nexthop, we do inner source IP lookup
to identify if it is ECMP source. If so, we force the flow processing.
Incase of control node peering with two MX, the BMS subnet would be
advertised by both MX leading to ECMP source . This is forcing the vxlan
tagged packets to be trapped to AGent with nh_id as VRF translate
nexthop. This leads to Agent marking the flow as short flow as Agent
already has set this flow with l3_nhid.

To avoid this scenario, the source lookup is done only for INET packets

Change-Id: Ic63c05ade2c7ee3757003d6cace8595eb015dc7a
closes-bug: #1467273