Nov overwriting VM's Mac in L2 encap nh
In R2.1, while supporting Floating IP, Agent on Private Network side
used to advertise Evpn route of Floating IP by stitching to Vhost's MAc.
Because of this to the reverse flow packets come to to compute node with
Vhosts mac and this mac is converted to VM's MAC, by l2 rewrite data.
But this overwrite is no more necessary as Agent started stiching this
IP with VM's MAC. So removing this rewrite data for L2 encap NH.

Change-Id: I375296662c5c1c4027e4c1312fdcdd9a8cb942a1
closes-bug: #1432754