Not overwriting VM's Mac in L2 encap nh
In R2.1, while supporting Floating IP, Agent on Private Network side
used to advertise Evpn route of Floating IP by stitching to Vhost's MAc.
Because of this to the reverse flow packets come to to compute node with
Vhosts mac and this mac is converted to VM's MAC, by l2 rewrite data.
But this overwrite is no more necessary as Agent started stiching this
IP with VM's MAC. So removing this rewrite data for L2 encap NH.

Change-Id: I5bf35524a58fa7acff9c1088643a5241a775fc50
closes-bug: #1432754