Handle out of order fragments
vRouter, as of now, drops fragments if the head of a fragmented packet
has not arrived yet, since it doesn't know the layer 4 port information.
To handle such scenarios, vRouter now caches such packets till head
of the fragmented packet arrives.

Datapath enqueues each fragment to a per-cpu queue of fragments. The
'assembler' dequeues the enqueued packets from the per-cpu queue and
enqueues such fragments in a packet cache. We have two separate
entities here because we wouldn't like to hold a lock in the datapath.
Once the head of the fragment arrives, the assembler flushes all the
held packets. There is also an aging timer that periodically scans for
inactive entries and frees them. This is more for packet loss cases
where some fragments don't make to the vrouter.

Change-Id: I7cac445f0674c05496ad7cf1c0eff5e2fed3b6ed
Closes-BUG: #1440230