Remove unnecessary UDP header manipulations in checksum functions
The dp-core is responsible for setting correct length of UDP packet and
zeroing the checksum field. It is not necessary to repeat this here before
sending the packet. Also the pkt_len() used before inserted length of the
entire packet (along with outer IP header + ether header), which was wrong.

Change-Id: I44f8855934c4b055af7b7549b5dec6ae5ebebe07