DPDK: Fix hashing of fragmented packets
We can not hash L4 ports as fragments other than the first have no L4 headers.
In result fragments of the same packets may have been distributed onto
different lcores and potentially be routed out of order, which is no supported
in the vRouter at the moment.

That caused random drops of packets in the situation when the VM sends already
fragmented packets to the vRouter.

Closes-Bug: #1474432
Change-Id: I7433819d00b33f3e0ab717552469c1d21183a0f7