Make deletion of nexthop an asynchronous operation
Currently, deletion of nexthop happens in the context of the vrouter
agent as a synchronous operation. To prevent traffic from accessing
invalid memory (that was a nexthop before deletion), we delay the
memory release using RCUs - the synchronous flavor of RCU.

In a scaled setup the synchronous RCU version can result in a
significant time spent doing delays, since for each delete the process
will delete, wait and then free the memory.

To get more useful work done in the time that is spent on wait, we
will make use of the asynchronous RCU, which will free agent to do
other work.

Change-Id: Ibbbef8bfeaaf40d3e087820d8121f99f9430f63b
Partial-BUG: #1465195