DPDK: Fix netlink attribute length of a response
vr_message_len that is used later in this function to set nla_len field of the
nlattr structure should not be aligned here as it leads to errors when parsing
sandesh messages on the other end (like dropstats, nh, etc. tools) due to wrong
length that may be bigger than actual message (i.e. length of 30 after
alignment is 32).

According to NetLink:

 <------- NLA_HDRLEN ------> <-- NLA_ALIGN(payload)-->
+---------------------+- - -+- - - - - - - - - -+- - -+
|        Header       | Pad |     Payload       | Pad |
|   (struct nlattr)   | ing |                   | ing |
+---------------------+- - -+- - - - - - - - - -+- - -+
 <-------------- nlattr->nla_len -------------->

This means that nla_len should not include the payload padding.

NOTE: Please note that currently the vRouter does not stick to NetLink in
100%. The nla_len in vRouter messages does not include the NLA_HDRLEN as it
hsould.

Change-Id: I675ae17b2cdbe50cb527c68d451b746fc7a06f7c
Partial-Bug: #1476745