Forcing L3 processing for Ecmp
When an ARP request is received from a VM which is part of ECMP to a non
ECMP destination, ARP response is sent using the stitching
information of destination route. This leads to VM learning  destination
VM's MAC directly enabling an L2 switching. But as source VM is part of
ECMP, the packets need to be L3 routed, as there is no ECMP support for L2
switched packet and also the reverse path packets (from non ECMP to ECMP)
are as such L3 routed.
To support this, ARP response is sent now with VRRP mac.

Change-Id: I595c0f14f3e09be31354d09af7522d750c1ca07f
closes-bug: #1472796