DPDK: Full VLAN support.
When vRouter works in VLAN, it expects packets tagged with VLAN tag
specified in the --vlan_tci command line parameter. Packets with
a matching tag will have the tag stripped and will be processed as
usual.

During a start up process, if --vlan_tci is specified, vRouter will
create additional KNI interface. Its name may be set using the
--vlan_fwd_intf_name command line parameter. Packets with a different
VLAN tag (or with no tag) will be forwarded to this interface. On top
of this VLAN forwarding interface subinterfaces can be created using
regular Linux tools, like vconfig.

If any packet is sent to the VLAN forwarding interface (or one of
its subinterfaces), it will be immediately forwarded to the wire
through the physical interface.

New dropstat counters has been added to report failures in enqueuing
packets to the ring and TXing them to the VLAN forwarding interface.

Additionaly, checksum offload for VLAN packets has been fixed and
hardware VLAN tag insertion has been enabled, if NIC supports that.

Closes-bug: #1484020
Closes-bug: #1484021
Closes-bug: #1485617
Change-Id: I8be6c577f84f8946ae514fd75b5654f9cbcc59c8