DPDK: Full VLAN support.
When vRouter works in VLAN, it expects packets tagged with VLAN tag
specified in the --vlan_tci command line parameter. Packets with
a matching tag will have it stripped and will be processed as usual.

During a start up process, if --vlan_tci is specified, vRouter will
create additional KNI interface. It's name may be set using the
--vlan_name command line parameter. Packets with a different VLAN
tag (or with no tag) will be forwarded to this interface. On top
of this VLAN forwarding interface subinterfaces can be created,
using regular Linux tools, like vconfig.

If any packet is sent to the VLAN forwarding interface (or one of
its subinterfaces), it will be immediately forwarded on the wire
through the physical interface.

Additionaly, checksum offload for VLAN packets has been fixed.

Closes-bug: #1484020
Closes-bug: #1484021
Change-Id: I8be6c577f84f8946ae514fd75b5654f9cbcc59c8